requirements.txt
pandas
numpy
matplotlib
seaborn
jupyter
sqlalchemy
psycopg2-binary
requests
dbt-core
dbt-postgres
dbt